Align table with database

+++ backend/src/main/java/com/sote/FarmRa/model/SensorNode.java
@@ -66,8 +66,8 @@ public class SensorNode {
     private Integer savedRssi;
     private Integer savedPacketLoss;
     private Integer savedBattery;
-    private Float savedTemperature;
-    private Float savedMoisture;
+    private Integer savedTemperature;
+    private Integer savedMoisture;
     private Boolean savedLight;
 
     @Setter
